requirements.txt
openai
tiktoken
python-dotenv
pinecone
pypdf
unstructured[pdf]